asignamos campo rol al formulario de registro, protegimos las rutas para crear usuario

+++ app/Http/Controllers/RoleController.php
@@ -3,21 +3,27 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
+use Spatie\Permission\Models\Role;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+
 
 class RoleController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
-        $texto = $request->input('texto'); //variable de el texto de la busqueda
-        $registros=Role::with('permissions')->where('name', 'like',"%{$texto}%")
-                        ->orderBy('id', 'desc')
-                        ->paginate(2);
+        $this->authorize('rol-list'); 
+        $texto = $request->input('texto'); // variable del texto de busqueda
+        $registros = Role::with('permissions')->where('name', 'like', "%{$texto}%")
+            ->orderBy('id', 'desc')
+            ->paginate(10);
+
         return view('role.index', compact('registros', 'texto'));
+
     }
 
     /**
@@ -25,6 +31,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $this->authorize('rol-create');
         $permissions = Permission::all();
         return view('role.action', compact('permissions'));
     }
@@ -34,14 +41,15 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('rol-create');
         $request->validate([
             'name' => 'required|unique:roles,name',
-            'permissions' => 'required|array',    
+            'permissions' => 'required|array',
         ]);
-        $registro = Roles::create(['name' => $request->name]);
+        $registro = Role::create(['name' => $request->name]);
         $registro->syncPermissions($request->permissions);
 
-        return redirect()->route('roles.index')-with('mensaje', 'Rol '.$registro->name.' creado satisfactoriamente');
+        return redirect()->route('roles.index')->with('mensaje', 'Rol '.$registro->name.' creado satisfactoriamente');
     }
 
     /**
@@ -57,9 +65,9 @@ class RoleController extends Controller
      */
     public function edit(string $id)
     {
+        $this->authorize('rol-edit');
         $registro = Role::findOrFail($id);
         $permissions = Permission::all();
-
         return view('role.action', compact('registro', 'permissions'));
     }
 
@@ -68,13 +76,15 @@ class RoleController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $this->authorize('rol-edit');
         $registro = Role::findOrFail($id);
         $request->validate([
-            'name' => 'required|unique:roles,name'.$registro->id,
-            'permissions' => 'required|array',    
+            'name' => 'required|unique:roles,name,'.$registro->id,
+            'permissions' => 'required|array',
         ]);
         $registro->update(['name' => $request->name]);
         $registro->syncPermissions($request->permissions);
+        return redirect()->route('roles.index')->with('mensaje', 'Registro '.$registro->name. '  actualizado correctamente');
     }
 
     /**
@@ -82,9 +92,10 @@ class RoleController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('rol-delete');
         $registro = Role::findOrFail($id);
         $registro->delete();
 
-        return redirect()->route('roles.index')->with('mensaje', $registro->name.' eliminado satisfactoriamente');
+        return redirect()->route('roles.index')->with('mensaje', $registro->name.' eliminado satisfactoriamente.');
     }
 }
