+++ app/Http/Controllers/UserController.php
@@ -1,20 +1,24 @@
 <?php
 
 namespace App\Http\Controllers;
-use App\Http\Requests\UserRequest;
+
+use App\Http\Requests\UserRequest; 
 use App\Models\User;
 use Illuminate\Http\Request;
-
+use Spatie\Permission\Models\Role;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class UserController extends Controller
 {
+    use AuthorizesRequests;
     /**
      * Display a listing of the resource.
      */
     public function index(Request $request)
     {
+        $this->authorize('user-list');
         $texto =$request->input('texto');
-        $registros=User::where('name','like','%'.$texto.'%')
+        $registros=User::with('roles')->where('name','like','%'.$texto.'%')
             ->orWhere('email','like','%'.$texto.'%')
             ->orderBy('id','desc')
             ->paginate(10);
@@ -27,7 +31,10 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('usuario.action');
+        $this->authorize('user-create');
+
+        $roles=Role::all();
+        return view('usuario.action', compact('roles'));
     }
 
     /**
@@ -35,13 +42,16 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        $this->authorize('user-create');
         $registro=new User();
         $registro->name=$request->input('name');
         $registro->email=$request->input('email');
         $registro->password=bcrypt($request->input('password'));
         $registro->activo=$request->input('activo');
         $registro->save();
-        return redirect()->route('usuarios.index')->with('mensaje','registro');
+
+        $registro->assignRole($request->input('role'));
+        return redirect()->route('usuarios.index')->with('mensaje','Registro '.$registro->name. ' agregado correctamente');
     }
 
     /**
@@ -57,8 +67,10 @@ class UserController extends Controller
      */
     public function edit($id)
     {
+        $this->authorize('user-edit');
         $registro=User::findOrFail($id);
-        return view('usuario.action', compact('registro'));
+        $roles=Role::all();
+        return view('usuario.action', compact('registro', 'roles'));
     }
 
     /**
@@ -66,13 +78,17 @@ class UserController extends Controller
      */
     public function update(UserRequest $request, $id)
     {
+        $this->authorize('user-edit');
         $registro=User::findOrFail($id);
         $registro->name=$request->input('name');
         $registro->email=$request->input('email');
         $registro->password=bcrypt($request->input('password'));
         $registro->activo=$request->input('activo');
         $registro->save();
-        return redirect()->route('usuarios.index')->with('mensaje','registro'.$registro->name.' Actualizado satisfatoriamente ' );
+
+        $registro->syncRoles([$request->input('role')]);
+
+        return redirect()->route('usuarios.index')->with('mensaje','Registro '.$registro->name.' actualizado satisfatoriamente. ' );
     }
 
     /**
@@ -80,12 +96,14 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('user-delete');
         $registro=User::findOrFail($id);
         $registro->delete();
         return redirect()->route('usuarios.index')->with('mensaje', $registro->name. ' eliminado satisfatoriamente.');
     }
 
     public function toggleStatus(User $usuario){
+        $this->authorize('user-activate');
         $usuario->activo = !$usuario->activo;
         $usuario->save();
         return redirect()->route('usuarios.index')->with('mensaje', 'El estado del usuario fue actualizado satisfactoriamente');
